+++ lib/glimmer-dsl-web/samples/hello/hello_style.rb
@@ -22,10 +22,20 @@
 require 'glimmer-dsl-web'
 
 class ButtonModel
-  attr_accessor :text, :pushed
+  WIDTH_MIN = 160
+  WIDTH_MAX = 960
+  HEIGHT_MIN = 100
+  HEIGHT_MAX = 600
+  FONT_SIZE_MIN = 40
+  FONT_SIZE_MAX = 200
+  
+  attr_accessor :text, :pushed, :width, :height, :font_size
   
   def initialize
     @text = 'Push'
+    @width = WIDTH_MIN
+    @height = HEIGHT_MIN
+    @font_size = FONT_SIZE_MIN
   end
   
   def push
@@ -35,38 +45,63 @@ class ButtonModel
   def text
     pushed ? 'Pull' : 'Push'
   end
+  
+  def width=(value)
+    @width = value
+    self.font_size = @width/4 if @font_size > @width/4
+  end
+  
+  def height=(value)
+    @height = value
+    self.font_size = @height/2.5 if @font_size > @height/2.5
+  end
+  
+  def font_size=(value)
+    @font_size = value
+    self.width = @font_size*4 if @height < @font_size*4
+    self.height = @font_size*2.5 if @height < @font_size*2.5
+  end
 end
 
 class StyledButton
   include Glimmer::Web::Component
   
-  option :text
-  
-  before_render do
-    @button_model = ButtonModel.new
-  end
+  option :button_model
   
   markup {
     button {
-      class_name <= [@button_model, :pushed,
+      inner_text <= [button_model, :text, computed_by: :pushed]
+      
+      class_name <= [button_model, :pushed,
                       on_read: ->(pushed) { pushed ? 'pushed' : 'pulled' }
                     ]
-                    
-      inner_html <= [@button_model, :text, computed_by: :pushed]
+      
+      style <= [ button_model, :width,
+                 on_read: method(:button_style_value) # convert value on read before storing in style
+               ]
+      
+      style <= [ button_model, :height,
+                 on_read: method(:button_style_value) # convert value on read before storing in style
+               ]
+      
+      style <= [ button_model, :font_size,
+                 on_read: method(:button_style_value) # convert value on read before storing in style
+               ]
       
       onclick do
-        @button_model.push
+        button_model.push
       end
     }
   }
   
   style {'
     button {
-      font-size: 60px;
       font-family: Courrier New, Courrier;
       border-radius: 5px;
       border-width: 17px;
-      border-color: lightgrey;
+      border-color: #ACC7D5;
+      background-color: #ADD8E6;
+      margin: 5px;
     }
     
     button.pulled {
@@ -77,24 +112,63 @@ class StyledButton
       border-style: inset;
     }
   '}
+  
+  def button_style_value
+    "
+      width: #{button_model.width}px;
+      height: #{button_model.height}px;
+      font-size: #{button_model.font_size}px;
+    "
+  end
+end
+
+class StyledButtonRangeInput
+  include Glimmer::Web::Component
+  
+  option :button_model
+  option :property
+  option :property_min
+  option :property_max
+  
+  markup {
+    input(type: 'range', min: property_min, max: property_max) {
+      value <=> [button_model, property]
+    }
+  }
 end
 
 class HelloStyle
   include Glimmer::Web::Component
   
+  before_render do
+    @button_model = ButtonModel.new
+  end
+  
   markup {
     div(class: 'hello-style') {
-      styled_button(text: 'Push Me') {
+      div(class: 'form-row') {
+        label('Styled Button Width:', for: 'styled-button-width-input')
+        styled_button_range_input(button_model: @button_model, property: :width, property_min: ButtonModel::WIDTH_MIN, property_max: ButtonModel::WIDTH_MAX, id: 'styled-button-width-input')
       }
-      button('Remove') {
-        onclick { remove }
+      div(class: 'form-row') {
+        label('Styled Button Height:', for: 'styled-button-height-input')
+        styled_button_range_input(button_model: @button_model, property: :height, property_min: ButtonModel::HEIGHT_MIN, property_max: ButtonModel::HEIGHT_MAX, id: 'styled-button-height-input')
       }
+      div(class: 'form-row') {
+        label('Styled Button Font Size:', for: 'styled-button-font-size-input')
+        styled_button_range_input(button_model: @button_model, property: :font_size, property_min: ButtonModel::FONT_SIZE_MIN, property_max: ButtonModel::FONT_SIZE_MAX, id: 'styled-button-font-size-input')
+      }
+      styled_button(button_model: @button_model)
     }
   }
   
   style {'
     .hello-style {
       padding: 20px;
+    }
+    
+    .hello-style .form-row {
+      margin: 10px 0;
     }
   '}
 end